spx-knowledge-backend/requirements.txt
# CPU则使用以下版本
#torch==2.3.1
#torchaudio==2.3.1
#torchvision==0.18.1
# GPU则优先执行
#pip install --index-url https://download.pytorch.org/whl/cu121 torch==2.3.1+cu121 torchaudio==2.3.1+cu121 torchvision==0.18.1+cu121
accelerate==1.11.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.0
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
billiard==4.2.2
cachetools==6.2.1
camelot-py==0.10.1
cbor==1.0.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
clamd==1.0.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.4.0
Deprecated>=1.3.1  # 修复 yanked 版本问题，使用更新的版本
dill==0.4.0
dnspython==2.8.0
easyocr==1.7.2
ecdsa==0.19.1
effdet==0.4.1
email-validator==2.3.0
emoji==2.15.0
et_xmlfile==2.0.0
Events==0.5
fastapi==0.120.1
filelock==3.20.0
filetype==1.2.0
FlagEmbedding==1.2.11
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
ghostscript==0.8.1
google-api-core==2.28.1
google-auth==2.42.0
google-cloud-vision==3.11.0
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ijson==3.4.0.post0
imageio==2.37.0
inscriptis==2.6.0
intel-openmp==2021.4.0
ir_datasets==0.5.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
kombu==5.5.4
langdetect==1.0.9
lazy_loader==0.4
lxml==6.0.2
lz4==4.4.5
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
minio==7.2.18
mistune==3.1.4
mkl==2021.4.0
ml_dtypes==0.5.3
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.5
ninja==1.13.0
nltk==3.9.2
numpy==1.26.4
olefile==0.47
ollama==0.6.0
omegaconf==2.3.0
onnx==1.19.1
onnxruntime==1.23.2
open_clip_torch==3.2.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opensearch-py==3.0.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
peft==0.17.1
pi_heif==1.1.1
pikepdf==10.0.0
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyMuPDF==1.26.6
PyMySQL==1.1.2
pyparsing==3.2.5
pypdf==6.1.3
PyPDF2==2.12.1
pypdfium2==5.0.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2025.2.18
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
python_clamd==0.0.2.dev0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
redis==7.0.1
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.2
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.49.0
sympy==1.14.0
tabulate==0.9.0
tbb==2021.13.1
threadpoolctl==3.6.0
tifffile==2025.10.16
timm==1.0.21
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
trec-car-tools==2.6
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unlzw3==0.2.3
unstructured==0.18.15
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
wrapt==2.0.0
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zlib-state==0.1.10
